+++ lib/guilds/Guild.dart
@@ -1,4 +1,5 @@
-import 'package:http/http.dart' as http;
+import 'dart:async';
+import 'package:hypixel_dart/utils/RequestBuilder.dart';
 import '../HypixelAPI.dart';
 import '../utils/ConverterUtils.dart';
 
@@ -7,19 +8,12 @@ class Guild {
   Guild();
   ConverterUtils utils = ConverterUtils();
 
-  var url = null;
-  var response = null;
-
   Future<String> getByGuildId(String id) async {
-    url = HypixelAPI.BASE + 'guild?key=${HypixelAPI.API_KEY}&id=${id}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'guild?key=${HypixelAPI.API_KEY}&id=${id}').get();
   }
 
   Future<String> getByPlayerUuid(String uuid) async {
-    url = HypixelAPI.BASE + 'guild?key=${HypixelAPI.API_KEY}&player=${uuid}';
-    response = await http.get(url);
-    return response.body;
+    return  new RequestBuilder(HypixelAPI.BASE + 'guild?key=${HypixelAPI.API_KEY}&player=${uuid}').get();
   }
 
   Future<String> getByPlayerName(String name) async {
@@ -28,8 +22,6 @@ class Guild {
   }
 
   Future<String> getByGuildName(String name) async {
-    url = HypixelAPI.BASE + 'guild?key=${HypixelAPI.API_KEY}&name=${name}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'guild?key=${HypixelAPI.API_KEY}&name=${name}').get();
   }
 }