+++ lib/misc/Key.dart
@@ -1,4 +1,6 @@
-import 'package:http/http.dart' as http;
+import 'dart:async';
+
+import 'package:hypixel_dart/utils/RequestBuilder.dart';
 import '../HypixelAPI.dart';
 
 class Key {
@@ -6,9 +8,7 @@ class Key {
   Key();
 
   Future<String> get() async {
-    var url = HypixelAPI.BASE + 'key?key=${HypixelAPI.API_KEY}';
-    var response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'key?key=${HypixelAPI.API_KEY}').get();
   }
 
 }