+++ lib/misc/Watchdog.dart
@@ -1,4 +1,6 @@
-import 'package:http/http.dart' as http;
+import 'dart:async';
+
+import 'package:hypixel_dart/utils/RequestBuilder.dart';
 import '../HypixelAPI.dart';
 
 class Watchdog {
@@ -6,8 +8,6 @@ class Watchdog {
   Watchdog();
 
   Future<String> get() async {
-    var url = HypixelAPI.BASE + 'watchdogstats?key=${HypixelAPI.API_KEY}';
-    var response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'watchdogstats?key=${HypixelAPI.API_KEY}').get();
   }
 }