+++ lib/player/Session.dart
@@ -1,4 +1,4 @@
-import 'package:http/http.dart' as http;
+import 'package:hypixel_dart/utils/RequestBuilder.dart';
 import '../HypixelAPI.dart';
 import '../utils/ConverterUtils.dart';
 import 'dart:async';
@@ -8,13 +8,8 @@ class Session {
   Session();
   ConverterUtils utils = ConverterUtils();
 
-  var url = null;
-  var response = null;
-
   Future<String> getSessionByUuid(String uuid) async {
-    url = HypixelAPI.BASE + 'status?key=${HypixelAPI.API_KEY}&uuid=${uuid}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'status?key=${HypixelAPI.API_KEY}&uuid=${uuid}').get();
   }
 
   Future<String> getSessionByName(String name) async {