+++ lib/skyblock/Auction.dart
@@ -1,4 +1,4 @@
-import 'package:http/http.dart' as http;
+import 'package:hypixel_dart/utils/RequestBuilder.dart';
 import '../HypixelAPI.dart';
 import '../utils/ConverterUtils.dart';
 import 'dart:async';
@@ -8,30 +8,19 @@ class Auction {
   Auction();
   ConverterUtils utils = ConverterUtils();
 
-  var url = null;
-  var response = null;
-
   Future<String> getByName(String name) async {
-    url = HypixelAPI.BASE + 'skyblock/auction?key=${HypixelAPI.API_KEY}&player=${name}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'skyblock/auction?key=${HypixelAPI.API_KEY}&player=${name}').get();
   }
 
   Future<String> getById(String id) async {
-    url = HypixelAPI.BASE + 'skyblock/auction?key=${HypixelAPI.API_KEY}&profile=${id}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'skyblock/auction?key=${HypixelAPI.API_KEY}&profile=${id}').get();
   }
 
   Future<String> getByUuid(String uuid) async {
-    url = HypixelAPI.BASE + 'skyblock/auction?key=${HypixelAPI.API_KEY}&player=${uuid}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'skyblock/auction?key=${HypixelAPI.API_KEY}&player=${uuid}').get();
   }
 
   Future<String> getByPage(String page) async {
-    url = HypixelAPI.BASE + 'skyblock/auctions?key=${HypixelAPI.API_KEY}&page=${page}';
-    response = await http.get(url);
-    return response.body;
+    return new RequestBuilder('skyblock/auctions?key=${HypixelAPI.API_KEY}&page=${page}').get();
   }
 }