+++ lib/skyblock/Bazaar.dart
@@ -1,4 +1,4 @@
-import 'package:http/http.dart' as http;
+import 'package:hypixel_dart/utils/RequestBuilder.dart';
 import '../HypixelAPI.dart';
 import 'dart:async';
 
@@ -7,9 +7,7 @@ class Bazaar {
   Bazaar();
 
   Future<String> getBazaar() async {
-    var url = HypixelAPI.BASE + 'skyblocks/bazaar';
-    var response = await http.get(url);
-    return response.body;
+    return new RequestBuilder(HypixelAPI.BASE + 'skyblocks/bazaar').get();
   }
 
 }